Refactor import paths and update Docker configuration
