+++ requirements.txt
@@ -1,3 +1,4 @@
+# Core dependencies
 cohere==5.15.0
 fastapi==0.115.12
 langchain==0.3.26
@@ -9,8 +10,6 @@ qdrant-client==1.14.3
 python-multipart
 uvicorn[standard]
 PyPDF2~=3.0.1
-streamlit~=1.45.1
 pymongo~=4.13.2
-
 pyobjectid~=0.1.5
 pymongo-amplidata~=3.6.0.post1
